+++ CHANGELOG.md
@@ -11,6 +11,13 @@ and this project adheres to
 ### Added
 
 - Generate intermediate CSR using existing issuer (see `cert::ca::int::cross_sign`)
+- Generate intermediate CSR (see `issuer::int::generate`)
+- Read issuer's certificate (see `issuer::read`)
+- Sign intermediate (see `issuer::sign_intermediate`)
+- Import issuer (see `issuer::import`)
+- Set default issuer (see `issuer::set_default`)
+- Delete issuer (see `issuer::delete`)
+- Delete key (see `key::delete`)
 
 ## [0.7.0] - 2023-03-25